refactor: create class controller for connection ws
