+++ src/websocket_server/index.ts
@@ -1,4 +1,5 @@
-import { RawData, WebSocketServer } from "ws";
+import { WebSocketServer } from "ws";
+import { WebSocketController } from "./controllers/web-socket-controller";
 // const WS_PORT = Number(process.env.WS_PORT) || 3000;
 
 // export const wss = new WebSocketServer({ port: WS_PORT });
@@ -6,29 +7,32 @@ import { RawData, WebSocketServer } from "ws";
 export class Wss {
   constructor(public port: number) {
     this.wss = new WebSocketServer({ port });
+    this.wsController = new WebSocketController();
     this.connection();
   }
   public wss: WebSocketServer;
+  protected wsController: WebSocketController;
 
   private connection() {
-    this.wss.on("connection", function connection(ws) {
-      ws.on("error", console.error);
-
-      ws.on("message", function message(data: RawData) {
-        console.log("reserved: %s", data);
-        const requestObject = JSON.parse(JSON.stringify(data));
-        const responseObject = {
-          type: "reg",
-          data: JSON.stringify({
-            name: requestObject.name,
-            index: 1,
-            error: false,
-            errorText: "no error",
-          }),
-          id: 0,
-        };
-        ws.send(JSON.stringify(responseObject));
-      });
-    });
+    this.wss.on(
+      "connection",
+      this.wsController.connection
+      // ws.on("error", console.error);
+      // ws.on("message", function message(data: RawData) {
+      //   console.log("reserved: %s", data);
+      //   const requestObject = JSON.parse(JSON.stringify(data));
+      //   const responseObject = {
+      //     type: "reg",
+      //     data: JSON.stringify({
+      //       name: requestObject.name,
+      //       index: 1,
+      //       error: false,
+      //       errorText: "no error",
+      //     }),
+      //     id: 0,
+      //   };
+      //   ws.send(JSON.stringify(responseObject));
+      // });
+    );
   }
 }
